requirements.txt
requests>=2.31.0
websockets==12.0